package/fluent-bit: fix openssl dependency
Fix the following build failure raised since the addition of the package
in commit 8e7ec68302e5bb3d014f7a42f8822646df3b732e:

Makefile:576: *** libopenssl is in the dependency chain of fluent-bit that has added it to its _DEPENDENCIES variable without selecting it or depending on it from Config.in.  Stop.

Fixes:
 - http://autobuild.buildroot.org/results/9c3b8e19063f7587877fdb0598739b55abccb3e8

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>