boot/barebox: use $$ instead of $ where needed
Commit 47a48df167fe4dc85c525cc114bd9344b2ebcfcb ("boot/barebox: pass
required environment variables for reproducible build") caused the
build to generate a spurious error message due to the $(shell
... date -d @$(SOURCE_DATE_EPOCH)) command being executed even when
BR2_REPRODUCIBLE was disabled.

This is due to the fact that variable references must be done using $$
within macros. This commit fixes that for the two variables that we
are referencing in the code added in commit
47a48df167fe4dc85c525cc114bd9344b2ebcfcb.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>