package/qt6/qt6serialport: use DESTDIR instead of --prefix for installation
Even though it works, overriding --prefix at installation time is a
bit weird. In order to be more consistent with what is done with other
build systems, use DESTDIR instead at installation time.

Note that $(DESTDIR) comes in addition to the
-DCMAKE_INSTALL_PREFIX=/usr that is passed at configure time, so with
this commit, the files continue to be installed in $(STAGING_DIR)/usr
and $(TARGET_DIR)/usr as they should be.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>